# backend/uzum_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Uzum market seller openapi",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-seller.uzum.uz/api/seller-openapi/",
      "description": "OpenAPI для продавцов UZUM MARKET"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    }
  ],
...
  "paths": {
    "/v1/product/shop/{shopId}": {
      "get": {
        "summary": "Получение SKU по ID магазина",
        "description": "Этот метод позволяет получить список товаров и остатков, доступных в указанном магазине.",
        "operationId": "getShopProductsByShopId",
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "searchQuery",
            "description": "Запрос для поиска товаров по названию или описанию.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "sortBy",
            "description": "Критерий сортировки товаров. Доступные варианты: DEFAULT, ORDERS, PRICE, ID, ROI, CONVERSION, LEFTOVERS, CREATED_AND_TITLE.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DEFAULT",
              "enum": [
                "DEFAULT",
                "ORDERS",
                "PRICE",
                "ID",
                "ROI",
                "CONVERSION",
                "LEFTOVERS",
                "CREATED_AND_TITLE"
              ]
            }
          },
          {
            "name": "order",
            "description": "Направление сортировки: ASC (по возрастанию) или DESC (по убыванию).",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "size",
            "description": "Количество товаров, возвращаемых в одном запросе.",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "description": "Номер страницы для постраничного просмотра товаров.",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "shopId",
            "description": "ID магазина, для которого необходимо получить товары.",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "productRank",
            "description": "Категория товара, которую нужно отобразить. Возможные значения: A, B, C, N, D.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "A",
                "B",
                "C",
                "N",
                "D"
              ]
            }
          },
          {
            "name": "filter",
            "description": "Фильтр для отображения товаров. Доступные значения: ALL, ACTIVE, INACTIVE, WARNING, WITH_SKU, ARCHIVE, DEFECTED, WITHOUT_REQUIRED_FILTERS.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ALL",
              "enum": [
                "ALL",
                "ACTIVE",
                "INACTIVE",
                "WARNING",
                "WITH_SKU",
                "ARCHIVE",
                "DEFECTED",
                "WITHOUT_REQUIRED_FILTERS"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AllProducts"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/shops": {
      "get": {
        "summary": "Получение списка собственных магазинов.",
        "operationId": "getOwnedShops",
        "tags": [
          "Shop"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenericResponseObject"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Токен авторизации без префикса Bearer"
      }
    },
    "headers": {
      "X-RateLimit-Remaining": {
        "description": "Оставшееся количество запросов на ближайшую секунду",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Replenish-Rate": {
        "description": "Скорость пополнения лимита запросов",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Burst-Capacity": {
        "description": "Максимальное количество запросов, которые могут быть выполнены за одну секунду",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Requested-Tokens": {
        "description": "Количество токенов, запрашиваемых за текущий запрос",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Limit-Per-Day": {
        "description": "Лимит запросов в день",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Remaining-Per-Day": {
        "description": "Оставшееся количество запросов в день",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "schemas": {
...
      "GenericResponseObject": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "object",
            "description": "Данные ответа, содержит информацию об объекте, возвращаемом сервисом."
          },
          "errors": {
            "type": "array",
            "description": "Список ошибок, если запрос не был успешно выполнен.",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Время формирования ответа, формат даты и времени."
          },
          "trace": {
            "type": "string",
            "description": "Трассировка для отладки ошибок."
          },
          "error": {
            "type": "string",
            "description": "Сообщение об ошибке, если оно имеется.",
            "readOnly": true,
            "example": "null"
          }
        }
      },
...
      "ApiErrorResponse": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Код ошибки, который может помочь в определении причины."
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке, предназначенное для отображения пользователю."
          },
          "detailMessage": {
            "type": "string",
            "deprecated": true,
            "description": "Более подробное сообщение об ошибке. Возможно, устарело и больше не используется."
          },
          "payload": {
            "type": "object",
            "description": "Дополнительные данные, связанные с ошибкой."
          }
        }
      },
      "AllProducts": {
        "type": "object",
        "properties": {
          "productList": {
            "type": "array",
            "description": "Список всех продуктов, доступных в магазине.",
            "items": {
              "$ref": "#/components/schemas/SellerProductCard"
            }
          },
          "totalProductsAmount": {
            "type": "integer",
            "format": "int32",
            "description": "Общее количество доступных продуктов."
          }
        }
      },
      "SellerProductCard": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "description": "Уникальный идентификатор продукта."
          },
          "category": {
            "type": "string",
            "description": "Категория, к которой принадлежит продукт."
          },
          "rating": {
            "type": "string",
            "description": "Рейтинг продукта, выраженный числом или символом."
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          },
          "moderationStatus": {
            "$ref": "#/components/schemas/ProductModerationStatus"
          },
          "commission": {
            "type": "number",
            "description": "Процент комиссии за продажу продукта. Возможно устарело и заменено на 'commissionDto'.",
            "format": "double",
            "deprecated": true
          },
          "commissionDto": {
            "$ref": "#/components/schemas/Commission"
          },
          "skuList": {
            "type": "array",
            "description": "Список SKU, связанных с продуктом.",
            "items": {
              "$ref": "#/components/schemas/SkuForTable"
            }
          },
          "skuTitle": {
            "type": "string",
            "description": "Название SKU продукта."
          },
          "image": {
            "type": "string",
            "description": "URL изображения продукта."
          },
          "previewImg": {
            "type": "string",
            "description": "URL предпросмотра изображения продукта."
          },
          "title": {
            "type": "string",
            "description": "Название продукта."
          },
          "quantityActive": {
            "type": "integer",
            "format": "int32",
            "description": "Количество активных единиц продукта на складе."
          },
          "quantityFbs": {
            "type": "integer",
            "format": "int32",
            "description": "Количество единиц, доступных в режиме FBS (Fulfillment by Seller)."
          }
        }
      },
      "Commission": {
        "type": "object",
        "description": "Детали комиссии, взимаемой за продукт, включая минимальную и максимальную ставку.",
        "properties": {
          "minCommission": {
            "type": "number",
            "format": "double",
            "description": "Минимальная ставка комиссии, взимаемая с продавца за продажу продукта."
          },
          "maxCommission": {
            "type": "number",
            "format": "double",
            "description": "Максимальная ставка комиссии, взимаемая с продавца за продажу продукта."
          }
        }
      },
...
      "SkuForTable": {
        "type": "object",
        "properties": {
          "skuTitle": {
            "type": "string",
            "description": "Название SKU (единицы хранения) товара."
          },
          "skuFullTitle": {
            "type": "string",
            "description": "Полное название SKU товара, включающее все детали."
          },
          "productTitle": {
            "type": "string",
            "description": "Название продукта, к которому относится SKU."
          },
          "skuId": {
            "type": "integer",
            "description": "Уникальный идентификатор SKU товара.",
            "format": "int64"
          },
          "quantityCreated": {
            "type": "integer",
            "description": "Общее количество созданных единиц SKU.",
            "format": "int32"
          },
          "quantityActive": {
            "type": "integer",
            "description": "Количество активных единиц SKU на складе.",
            "format": "int32"
          },
          "quantityFbs": {
            "type": "integer",
            "description": "Количество единиц SKU, доступных в режиме FBS (Fulfillment by Seller).",
            "format": "int32"
          },
          "quantityAdditional": {
            "type": "integer",
            "description": "Дополнительное количество SKU, которое может быть задействовано при необходимости.",
            "format": "int32"
          },
          "quantityOnPhotoStudio": {
            "type": "integer",
            "description": "Количество SKU, находящихся в фотостудии для съемки.",
            "format": "int32"
          },
          "quantityArchived": {
            "type": "integer",
            "description": "Количество архивированных SKU.",
            "format": "int32"
          },
          "quantitySold": {
            "type": "integer",
            "description": "Общее количество проданных SKU.",
            "format": "int32"
          },
          "quantityReturned": {
            "type": "integer",
            "description": "Количество возвращенных покупателем SKU.",
            "format": "int32"
          },
          "quantityMissing": {
            "type": "integer",
            "description": "Количество недостающих SKU (параметр устарел).",
            "format": "int32"
          },
          "quantityDefected": {
            "type": "integer",
            "description": "Количество дефектных SKU.",
            "format": "int32"
          },
          "quantityPending": {
            "type": "integer",
            "description": "Количество SKU в ожидании обработки или отправки.",
            "format": "int32"
          },
          "returnedPercentage": {
            "type": "number",
            "description": "Процент возвращенных SKU относительно общего количества проданных.",
            "format": "double"
          },
          "barcode": {
            "type": "integer",
            "description": "Штрих-код SKU для идентификации.",
            "format": "int64"
          },
          "archived": {
            "type": "boolean",
            "description": "Показывает, находится ли SKU в архиве."
          },
          "commission": {
            "type": "number",
            "description": "Комиссия, связанная с продажей данного SKU.",
            "format": "double"
          },
          "characteristics": {
            "type": "string",
            "description": "Описание характеристик SKU, может включать основные атрибуты."
          },
          "purchasePrice": {
            "type": "integer",
            "description": "Цена закупки данного SKU.",
            "format": "int64"
          },
          "price": {
            "type": "integer",
            "description": "Цена продажи SKU.",
            "format": "int64"
          },
          "blockingReason": {
            "type": "string",
            "description": "Причина блокировки SKU, если оно заблокировано."
          },
          "skuBlockReason": {
            "$ref": "#/components/schemas/SkuBlockReason"
          },
          "blocked": {
            "type": "boolean",
            "description": "Показывает, заблокировано ли SKU."
          },
          "rankInfo": {
            "$ref": "#/components/schemas/ProductRank"
          },
          "article": {
            "type": "string",
            "description": "Артикул SKU, используемый для идентификации продукта продавцом."
          },
          "turnover": {
            "type": "number",
            "description": "Оборот по данному SKU в денежном выражении.",
            "format": "double"
          },
          "dimensionalGroup": {
            "type": "string",
            "description": "Габаритная группа SKU, указывающая его размер или вес."
          },
          "paidStorageDimensionalGroup": {
            "description": "Габаритная группа для оплачиваемого хранения SKU.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DimensionalGroup"
              }
            ]
          },
          "paidStoragePriceItem": {
            "type": "integer",
            "description": "Стоимость единицы хранения для SKU в оплачиваемом режиме.",
            "format": "int64"
          },
          "paidStorageAmount": {
            "type": "integer",
            "description": "Общая стоимость хранения SKU в оплачиваемом режиме.",
            "format": "int64"
          },
          "actualDimensionalGroup": {
            "description": "Фактическая габаритная группа SKU.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DimensionalGroup"
              }
            ]
          },
          "hasStudioPhoto": {
            "type": "boolean",
            "description": "Показывает, есть ли у SKU студийное фото."
          },
          "avgdsales": {
            "type": "number",
            "description": "Среднее количество продаж данного SKU в день.",
            "format": "double"
          },
          "avgdquantity": {
            "type": "number",
            "description": "Среднее количество активных единиц SKU в день.",
            "format": "double"
          },
          "pstorage": {
            "type": "boolean",
            "description": "Показывает, находится ли SKU на платном складе хранения."
          },
          "previewImage": {
            "type": "string",
            "description": "Ссылка на фото превью товара"
          },
          "ikpu": {
            "type": "string",
            "description": "Идентификационный код продукции и услуг для Узбекистана, представлен в виде набора символов (чисел)"
          },
          "sellerItemCode": {
              "type": "string",
              "description": "Идентификатор селлера"
          }
        }
      },
      "ProductRank": {
        "type": "object",
        "description": "Информация о ранге продукта, включая его позицию на рынке.",
        "properties": {
          "rank": {
            "type": "string",
            "description": "Ранг продукта, представляющий его классификацию по уровню популярности или качества.",
            "enum": [
              "A",
              "B",
              "C",
              "N",
              "D"
            ]
          },
          "rankValue": {
            "type": "string",
            "description": "Числовое значение ранга, характеризующее позицию продукта."
          },
          "dateUpdated": {
            "type": "string",
            "description": "Дата последнего обновления информации о ранге.",
            "format": "date-time"
          }
        }
      },
...
      "ProductStatus": {
        "type": "object",
        "description": "Статус продукта, показывающий текущее состояние товара.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Уникальный идентификатор статуса продукта."
          },
          "title": {
            "type": "string",
            "description": "Название статуса продукта, например, 'Готов к отправке'."
          },
          "value": {
            "type": "string",
            "enum": [
              "NO_SKU",
              "IN_STOCK",
              "READY_TO_SEND",
              "SENT",
              "RUN_OUT",
              "BLOCKED",
              "SKU_BLOCKED",
              "ARCHIVED",
              "DELETED",
              "PERM_BANNED",
              "NOT_READY_TO_SEND"
            ],
            "description": "Значение статуса, указывающее текущее состояние продукта."
          },
          "description": {
            "type": "string",
            "description": "Описание статуса, объясняющее его значение для продавца."
          },
          "color": {
            "type": "string",
            "description": "Цвет, ассоциированный с данным статусом для визуальной индикации."
          },
          "additional": {
            "type": "array",
            "description": "Дополнительные статусы продукта, например, временные состояния.",
            "items": {
              "$ref": "#/components/schemas/ProductAdditionalStatus"
            }
          }
        }
      },
      "ProductAdditionalStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Уникальный идентификатор дополнительного статуса продукта."
          },
          "title": {
            "type": "string",
            "description": "Название дополнительного статуса."
          },
          "description": {
            "type": "string",
            "description": "Описание дополнительного статуса, объясняющее его значение."
          },
          "color": {
            "type": "string",
            "description": "Цвет, ассоциированный с дополнительным статусом, для визуальной индикации."
          },
          "icon": {
            "type": "string",
            "description": "Иконка, отображаемая для дополнительного статуса."
          }
        }
      },
      "ProductModerationStatus": {
        "type": "object",
        "description": "Статус модерации, указывающий, прошел ли продукт проверку.",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название статуса модерации, например, 'На модерации'."
          },
          "color": {
            "type": "string",
            "description": "Цвет, связанный со статусом модерации, для визуального отображения."
          },
          "value": {
            "type": "string",
            "enum": [
              "NOT_MODERATED",
              "MODERATED",
              "ON_MODERATION",
              "HAS_COMPLAINTS",
              "ON_PREMODERATION",
              "PERM_BANNED"
            ],
            "description": "Значение статуса модерации, показывающее текущий этап модерации продукта."
          },
          "hasAnySkuBlocked": {
            "type": "boolean",
            "description": "Показывает, был ли заблокирован какой-либо SKU из данного продукта."
          }
        }
      },
...
      "SkuBlockReason": {
        "type": "object",
        "description": "Детали причины блокировки SKU, если оно заблокировано.",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название причины блокировки SKU."
          },
          "message": {
            "type": "string",
            "description": "Сообщение с деталями причины блокировки."
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Дата, когда SKU был заблокирован."
          }
        }
      },
      "DimensionalGroup": {
        "type": "object",
        "properties": {
          "group": {
            "type": "string",
            "enum": [
              "SMALL",
              "MEDIUM",
              "LARGE",
              "UNKNOWN"
            ],
            "description": "Категория размера товара, например, малая, средняя или большая."
          },
          "title": {
            "type": "string",
            "description": "Название категории размера."
          }
        }
      },
...
      "OrganizationDto": {
        "type": "object",
        "description": "Информация по магазину",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Уникальный идентификатор магазина",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "Название магазина"
          }
        }
      },
...
    }
  }
}